Overhauled music system
A new method has been added: it generates a music wav file containing the 'tune'. This is played then deleted.
This is now the default version for Windows 7+, Linux and MacOS (Darwin kernel). Winsound kept for Windows XP (not tested though).
Known bug: wave module only loads half of the file, therefore the file is doubled before writing: this may cause the tune to double if the wave is fixed. Suggested: Implement own version
Numpy is now required for sound.